separate utilities from signal decomposition problems
